pyproject.toml
[project]
name = "pyobs-gui"
version = "1.4.1"
description = "A remote GUI for pyobs"
authors = [{ name = "Tim-Oliver Husser", email = "thusser@uni-goettingen.de" }]
requires-python = ">=3.11,<3.14"
license = "MIT"
dependencies = [
    "colour>=0.1.5,<0.2",
    "matplotlib>=3.10.1,<4",
    "qasync>=0.27.1,<0.28",
    "sunpy>=6.1.1,<7",
    "QtAwesome>=1.4.0,<2",
    "qfitswidget>=0.11.0",
    "pyobs-core>=1.24.0",
    "pyside6>=6.10.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0,<26",
    "flake8>=7.3.0",
    "matplotlib-stubs>=0.1.0",
    "mypy>=1.17.0",
    "pandas-stubs>=2.3.0.250703",
    "pre-commit>=4.2.0,<5",
    "pyside6-stubs>=6.7.3.0",
    "sphinx-rtd-theme>=3.0.2,<4",
    "Sphinx>=8.2.3,<9",
    "types-defusedxml>=0.7.0.20250708",
    "types-python-dateutil>=2.9.0.20250708",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 120
target-version = ['py311']
